Revert "Clear AddComponent function (#75)"
This reverts commit d30d6917cbc35f590c6d18ea521c0058af8f2d5c.

+++ Breakout/Script/GameLevel.h
@@ -11,10 +11,10 @@ class GameLevel
 {
 private:
     const std::string fileName;
-    size_t levelHeight;
-    size_t levelWidth;
+    unsigned int levelHeight;
+    unsigned int levelWidth;
 
-    void Init(const std::vector<std::vector<size_t>>& brickData, bool areBricksActive);
+    void Init(const std::vector<std::vector<unsigned int>>& brickData, bool areBricksActive);
     void CalcBrickSize(const Learning2DEngine::Render::Resolution& resolution);
 public:
     std::vector<BrickController*> bricks;
